backend/requirements.txt
celery==4.4.7
eventlet==0.30.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==5.0.1
flower==0.9.5
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
hiredis==1.1.0
librabbitmq==2.0.0
kombu==4.6.11
prometheus-client==0.8.0
python-socketio==5.0.4
PyYAML==5.3.1
psycopg2==2.8.6
redis==3.5.3
